Add test page for javascript-info ifelse tutorial
adds a page I spun up to type in my javascript answers for the bottom of
the page. The html boilerplate auto-generated by Atom was nice, and I am
enjoying it so far, the only thing I had to type was 'doc' to get the
DOCTYPE segment and '<ht' to get the rest to autogenerate. *neat.gif*