Add Rock Paper Scissors Project
This is the file for the rock paper scissors project. It's done enough
for this online course and goes beyond the parameters set by the lesson.
I would like to have the results be a table after playing and have a
different way of interfacing with the user besides using the prompts...