Add JS exercises from js basics port 4
This adds the code borrowed from the git repo for javasript excercises at the
Odin Project's official repo. adding to mine for ease of commits and such. I
didn't want to start a new repo for just some example code that I am barely
modifying but also want to not lose my work. It is the recommended study at
the end of js-fundamentals-part4 so I am sticking it in that directory.
And there's nothing you can do to stop me. muahahahaha