Starts creating R.P.S. gui for Dom Manip section
the gui is started, but it is bed time on a work night. :/ more to do
tomorrow. good news is the chimney has some siding on it and the chase
cap I hated doesn't fit so I have no choice but to get a better one made
which helps my house look better for sure. also got some yoga and
laundry in. huzzah!